attach: use function dispatcher
