+++ attach/dlg_attach.c
@@ -76,12 +76,14 @@
 #include "email/lib.h"
 #include "core/lib.h"
 #include "gui/lib.h"
+#include "index/lib.h"
 #include "menu/lib.h"
 #include "ncrypt/lib.h"
 #include "question/lib.h"
 #include "send/lib.h"
 #include "attachments.h"
 #include "format_flags.h"
+#include "functions.h"
 #include "hdrline.h"
 #include "hook.h"
 #include "mutt_attach.h"
@@ -103,42 +105,6 @@ static const struct Mapping AttachHelp[] = {
   { NULL, 0 },
   // clang-format on
 };
-
-/**
- * attach_collapse - Close the tree of the current attachment
- * @param actx Attachment context
- * @param menu Menu listing Attachments
- */
-void attach_collapse(struct AttachCtx *actx, struct Menu *menu)
-{
-  int rindex, curlevel;
-
-  struct AttachPtr *cur_att = current_attachment(actx, menu);
-  cur_att->collapsed = !cur_att->collapsed;
-  /* When expanding, expand all the children too */
-  if (cur_att->collapsed)
-    return;
-
-  curlevel = cur_att->level;
-  const int index = menu_get_index(menu);
-  rindex = actx->v2r[index] + 1;
-
-  const bool c_digest_collapse =
-      cs_subset_bool(NeoMutt->sub, "digest_collapse");
-  while ((rindex < actx->idxlen) && (actx->idx[rindex]->level > curlevel))
-  {
-    if (c_digest_collapse && (actx->idx[rindex]->body->type == TYPE_MULTIPART) &&
-        mutt_istr_equal(actx->idx[rindex]->body->subtype, "digest"))
-    {
-      actx->idx[rindex]->collapsed = true;
-    }
-    else
-    {
-      actx->idx[rindex]->collapsed = false;
-    }
-    rindex++;
-  }
-}
 
 /**
  * attach_config_observer - Notification that a Config Variable has changed - Implements ::observer_t - @ingroup observer_api
@@ -475,90 +441,6 @@ static int attach_window_observer(struct NotifyCallback *nc)
 }
 
 /**
- * check_attach - Check if in attach-message mode
- * @retval true Mailbox is readonly
- */
-bool check_attach(void)
-{
-  if (OptAttachMsg)
-  {
-    mutt_flushinp();
-    mutt_error(_("Function not permitted in attach-message mode"));
-    return true;
-  }
-
-  return false;
-}
-
-/**
- * check_readonly - Check if the Mailbox is readonly
- * @param m Mailbox
- * @retval true Mailbox is readonly
- */
-bool check_readonly(struct Mailbox *m)
-{
-  if (!m || m->readonly)
-  {
-    mutt_flushinp();
-    mutt_error(_("Mailbox is read-only"));
-    return true;
-  }
-
-  return false;
-}
-
-/**
- * recvattach_extract_pgp_keys - Extract PGP keys from attachments
- * @param actx Attachment context
- * @param menu Menu listing attachments
- */
-void recvattach_extract_pgp_keys(struct AttachCtx *actx, struct Menu *menu)
-{
-  if (!menu->tagprefix)
-  {
-    struct AttachPtr *cur_att = current_attachment(actx, menu);
-    crypt_pgp_extract_key_from_attachment(cur_att->fp, cur_att->body);
-  }
-  else
-  {
-    for (int i = 0; i < actx->idxlen; i++)
-    {
-      if (actx->idx[i]->body->tagged)
-      {
-        crypt_pgp_extract_key_from_attachment(actx->idx[i]->fp, actx->idx[i]->body);
-      }
-    }
-  }
-}
-
-/**
- * recvattach_pgp_check_traditional - Is the Attachment inline PGP?
- * @param actx Attachment to check
- * @param menu Menu listing Attachments
- * @retval 1 The (tagged) Attachment(s) are inline PGP
- *
- * @note If the menu->tagprefix is set, all the tagged attachments will be checked.
- */
-int recvattach_pgp_check_traditional(struct AttachCtx *actx, struct Menu *menu)
-{
-  int rc = 0;
-
-  if (!menu->tagprefix)
-  {
-    struct AttachPtr *cur_att = current_attachment(actx, menu);
-    rc = crypt_pgp_check_traditional(cur_att->fp, cur_att->body, true);
-  }
-  else
-  {
-    for (int i = 0; i < actx->idxlen; i++)
-      if (actx->idx[i]->body->tagged)
-        rc = rc || crypt_pgp_check_traditional(actx->idx[i]->fp, actx->idx[i]->body, true);
-  }
-
-  return rc;
-}
-
-/**
  * dlg_select_attachment - Show the attachments in a Menu
  * @param sub Config Subset
  * @param m   Mailbox
@@ -609,346 +491,17 @@ void dlg_select_attachment(struct ConfigSubset *sub, struct Mailbox *m,
     if (op == OP_NULL)
       op = menu_loop(menu);
     window_redraw(dlg);
-    switch (op)
+    int rc = attach_function_dispatcher(dlg, op);
+    if (rc == IR_DONE)
+      break;
+    if (rc == IR_CONTINUE)
     {
-      case OP_ATTACHMENT_VIEW_MAILCAP:
-      {
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_view_attachment(cur_att->fp, cur_att->body, MUTT_VA_MAILCAP,
-                             priv->actx->email, priv->actx, priv->menu->win);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        break;
-      }
-
-      case OP_ATTACHMENT_VIEW_TEXT:
-      {
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_view_attachment(cur_att->fp, cur_att->body, MUTT_VA_AS_TEXT,
-                             priv->actx->email, priv->actx, priv->menu->win);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        break;
-      }
-
-      case OP_ATTACHMENT_VIEW_PAGER:
-      {
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_view_attachment(cur_att->fp, cur_att->body, MUTT_VA_PAGER,
-                             priv->actx->email, priv->actx, priv->menu->win);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        break;
-      }
-
-      case OP_DISPLAY_HEADERS:
-      case OP_ATTACHMENT_VIEW:
-        op = mutt_attach_display_loop(priv->sub, priv->menu, op,
-                                      priv->actx->email, priv->actx, true);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        continue;
-
-      case OP_ATTACHMENT_COLLAPSE:
-      {
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        if (!cur_att->body->parts)
-        {
-          mutt_error(_("There are no subparts to show"));
-          break;
-        }
-        attach_collapse(priv->actx, priv->menu);
-        mutt_update_recvattach_menu(priv->actx, priv->menu, false);
-        break;
-      }
-
-      case OP_FORGET_PASSPHRASE:
-        crypt_forget_passphrase();
-        break;
-
-      case OP_EXTRACT_KEYS:
-        if (WithCrypto & APPLICATION_PGP)
-        {
-          recvattach_extract_pgp_keys(priv->actx, priv->menu);
-          menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        }
-        break;
-
-      case OP_CHECK_TRADITIONAL:
-        if (((WithCrypto & APPLICATION_PGP) != 0) &&
-            recvattach_pgp_check_traditional(priv->actx, priv->menu))
-        {
-          priv->actx->email->security = crypt_query(NULL);
-          menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        }
-        break;
-
-      case OP_ATTACHMENT_PRINT:
-      {
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_print_attachment_list(priv->actx, cur_att->fp,
-                                   priv->menu->tagprefix, cur_att->body);
-        break;
-      }
-
-      case OP_ATTACHMENT_PIPE:
-      {
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_pipe_attachment_list(priv->actx, cur_att->fp,
-                                  priv->menu->tagprefix, cur_att->body, false);
-        break;
-      }
-
-      case OP_ATTACHMENT_SAVE:
-      {
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_save_attachment_list(priv->actx, cur_att->fp, priv->menu->tagprefix,
-                                  cur_att->body, priv->actx->email, priv->menu);
-
-        const bool c_resolve = cs_subset_bool(NeoMutt->sub, "resolve");
-        const int index = menu_get_index(priv->menu) + 1;
-        if (!priv->menu->tagprefix && c_resolve && (index < priv->menu->max))
-          menu_set_index(priv->menu, index);
-        break;
-      }
-
-      case OP_ATTACHMENT_DELETE:
-        if (check_readonly(priv->mailbox))
-          break;
-
-#ifdef USE_POP
-        if (priv->mailbox->type == MUTT_POP)
-        {
-          mutt_flushinp();
-          mutt_error(_("Can't delete attachment from POP server"));
-          break;
-        }
-#endif
-
-#ifdef USE_NNTP
-        if (priv->mailbox->type == MUTT_NNTP)
-        {
-          mutt_flushinp();
-          mutt_error(_("Can't delete attachment from news server"));
-          break;
-        }
-#endif
-
-        if ((WithCrypto != 0) && (priv->actx->email->security & SEC_ENCRYPT))
-        {
-          mutt_message(_("Deletion of attachments from encrypted messages is "
-                         "unsupported"));
-          break;
-        }
-        if ((WithCrypto != 0) && (priv->actx->email->security & (SEC_SIGN | SEC_PARTSIGN)))
-        {
-          mutt_message(_("Deletion of attachments from signed messages may "
-                         "invalidate the signature"));
-        }
-        if (!priv->menu->tagprefix)
-        {
-          struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-          if (cur_att->parent_type == TYPE_MULTIPART)
-          {
-            cur_att->body->deleted = true;
-            const bool c_resolve = cs_subset_bool(NeoMutt->sub, "resolve");
-            const int index = menu_get_index(priv->menu) + 1;
-            if (c_resolve && (index < priv->menu->max))
-            {
-              menu_set_index(priv->menu, index);
-            }
-            else
-              menu_queue_redraw(priv->menu, MENU_REDRAW_CURRENT);
-          }
-          else
-          {
-            mutt_message(
-                _("Only deletion of multipart attachments is supported"));
-          }
-        }
-        else
-        {
-          for (int i = 0; i < priv->menu->max; i++)
-          {
-            if (priv->actx->idx[i]->body->tagged)
-            {
-              if (priv->actx->idx[i]->parent_type == TYPE_MULTIPART)
-              {
-                priv->actx->idx[i]->body->deleted = true;
-                menu_queue_redraw(priv->menu, MENU_REDRAW_INDEX);
-              }
-              else
-              {
-                mutt_message(
-                    _("Only deletion of multipart attachments is supported"));
-              }
-            }
-          }
-        }
-        break;
-
-      case OP_ATTACHMENT_UNDELETE:
-        if (check_readonly(priv->mailbox))
-          break;
-        if (!priv->menu->tagprefix)
-        {
-          struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-          cur_att->body->deleted = false;
-          const bool c_resolve = cs_subset_bool(NeoMutt->sub, "resolve");
-          const int index = menu_get_index(priv->menu) + 1;
-          if (c_resolve && (index < priv->menu->max))
-          {
-            menu_set_index(priv->menu, index);
-          }
-          else
-            menu_queue_redraw(priv->menu, MENU_REDRAW_CURRENT);
-        }
-        else
-        {
-          for (int i = 0; i < priv->menu->max; i++)
-          {
-            if (priv->actx->idx[i]->body->tagged)
-            {
-              priv->actx->idx[i]->body->deleted = false;
-              menu_queue_redraw(priv->menu, MENU_REDRAW_INDEX);
-            }
-          }
-        }
-        break;
-
-      case OP_RESEND:
-      {
-        if (check_attach())
-          break;
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_attach_resend(cur_att->fp, priv->mailbox, priv->actx,
-                           priv->menu->tagprefix ? NULL : cur_att->body);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        break;
-      }
-
-      case OP_BOUNCE_MESSAGE:
-      {
-        if (check_attach())
-          break;
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_attach_bounce(priv->mailbox, cur_att->fp, priv->actx,
-                           priv->menu->tagprefix ? NULL : cur_att->body);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        break;
-      }
-
-      case OP_FORWARD_MESSAGE:
-      {
-        if (check_attach())
-          break;
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_attach_forward(cur_att->fp, priv->actx->email, priv->actx,
-                            priv->menu->tagprefix ? NULL : cur_att->body, SEND_NO_FLAGS);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        break;
-      }
-
-#ifdef USE_NNTP
-      case OP_FORWARD_TO_GROUP:
-      {
-        if (check_attach())
-          break;
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_attach_forward(cur_att->fp, priv->actx->email, priv->actx,
-                            priv->menu->tagprefix ? NULL : cur_att->body, SEND_NEWS);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        break;
-      }
-
-      case OP_FOLLOWUP:
-      {
-        if (check_attach())
-          break;
-
-        const enum QuadOption c_followup_to_poster =
-            cs_subset_quad(NeoMutt->sub, "followup_to_poster");
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        if (!cur_att->body->email->env->followup_to ||
-            !mutt_istr_equal(cur_att->body->email->env->followup_to,
-                             "poster") ||
-            (query_quadoption(c_followup_to_poster,
-                              _("Reply by mail as poster prefers?")) != MUTT_YES))
-        {
-          mutt_attach_reply(cur_att->fp, priv->mailbox, priv->actx->email,
-                            priv->actx, priv->menu->tagprefix ? NULL : cur_att->body,
-                            SEND_NEWS | SEND_REPLY);
-          menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-          break;
-        }
-      }
-#endif
-      /* fallthrough */
-      case OP_REPLY:
-      case OP_GROUP_REPLY:
-      case OP_GROUP_CHAT_REPLY:
-      case OP_LIST_REPLY:
-      {
-        if (check_attach())
-          break;
-
-        SendFlags flags = SEND_REPLY;
-        if (op == OP_GROUP_REPLY)
-          flags |= SEND_GROUP_REPLY;
-        else if (op == OP_GROUP_CHAT_REPLY)
-          flags |= SEND_GROUP_CHAT_REPLY;
-        else if (op == OP_LIST_REPLY)
-          flags |= SEND_LIST_REPLY;
-
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_attach_reply(cur_att->fp, priv->mailbox, priv->actx->email, priv->actx,
-                          priv->menu->tagprefix ? NULL : cur_att->body, flags);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        break;
-      }
-
-      case OP_LIST_SUBSCRIBE:
-        if (!check_attach())
-          mutt_send_list_subscribe(priv->mailbox, priv->actx->email);
-        break;
-
-      case OP_LIST_UNSUBSCRIBE:
-        if (!check_attach())
-          mutt_send_list_unsubscribe(priv->mailbox, priv->actx->email);
-        break;
-
-      case OP_COMPOSE_TO_SENDER:
-      {
-        if (check_attach())
-          break;
-        struct AttachPtr *cur_att = current_attachment(priv->actx, priv->menu);
-        mutt_attach_mail_sender(priv->actx,
-                                priv->menu->tagprefix ? NULL : cur_att->body);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_FULL);
-        break;
-      }
-
-      case OP_ATTACHMENT_EDIT_TYPE:
-        recvattach_edit_content_type(priv->actx, priv->menu, priv->actx->email);
-        menu_queue_redraw(priv->menu, MENU_REDRAW_INDEX);
-        break;
-
-      case OP_EXIT:
-        priv->actx->email->attach_del = false;
-        for (int i = 0; i < priv->actx->idxlen; i++)
-        {
-          if (priv->actx->idx[i]->body && priv->actx->idx[i]->body->deleted)
-          {
-            priv->actx->email->attach_del = true;
-            break;
-          }
-        }
-        if (priv->actx->email->attach_del)
-          priv->actx->email->changed = true;
-
-        mutt_actx_free(&priv->actx);
-        goto done;
+      op = priv->op;
+      continue;
     }
 
     op = OP_NULL;
   }
 
-done:
   simple_dialog_free(&dlg);
 }
